ipasir: add ipasir_solve_final function
This function can be used to perform the last know call to the SAT
solver. A reason for the existence of this functoin is to provide a
simple way to integrate simplification into tools that use SAT
solvers as a backend. Most SAT solvers are still not capable of
incremental formula simplification. Hence, a simple way to be able
to use simplification is to provide the user with a method that
indicates that no further calls to the solver method will be issued.

Most simplification implementation of SAT solvers are not capable of
being executed in case the solve method has been called already.
Hence, SAT solvers implementing this function differently to the
ipasir_solve function might only use simplification if the call to
ipasir_solve_final is the very first (and single) call to a ipasir
solve function.