ipasir: add version control
The set of functions to be implemented changed over time. However,
it is not simple to determine whether a SAT backend fulfills the
criteria an application with the interface might have. To allow to
easily specify the needs, add a version number, which controls
which functions have to be implemented.

As there are multiple version available already, also wrap the new
functions with the version number, so that the version number can
be used in a backwards compatible way.